split up and refactored basic example
Simplified the basic example to work with a single table for
clarity. Added the example with two linked tables as
tables.hs. Added the multi-database support as multidb.hs.
