+++ examples/groundhog-examples.cabal
@@ -23,6 +23,26 @@ executable basic
                      , transformers
                      , groundhog-th
                      , groundhog-sqlite
+
+  default-language:    Haskell2010
+
+executable tables
+  main-is:             tables.hs
+  build-depends:       base >=4 && <5
+                     , groundhog
+                     , transformers
+                     , groundhog-th
+                     , groundhog-sqlite
+
+  default-language:    Haskell2010
+
+executable multidb
+  main-is:             multidb.hs
+  build-depends:       base >=4 && <5
+                     , groundhog
+                     , transformers
+                     , groundhog-th
+                     , groundhog-sqlite
                      , groundhog-mysql
                      , groundhog-postgresql
